Renamed build-logic to buildSrc and updated Gradle to 8.0
