+++ base/build.gradle.kts
@@ -1,10 +1,10 @@
 dependencies {
-    api(libs.checkerQual)
-    api(libs.cumulus)
-    api(libs.events) {
-        exclude(group = "com.google.guava", module = "guava")
-        exclude(group = "org.lanternpowered", module = "lmbda")
-    }
+  api(libs.checker.qual)
+  api(libs.cumulus)
+  api(libs.events) {
+    exclude(group = "com.google.guava", module = "guava")
+    exclude(group = "org.lanternpowered", module = "lmbda")
+  }
 }
 
 version = property("baseApiVersion")!!
