--- build-logic/build.gradle.kts
@@ -1,20 +1,0 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-plugins {
-    `kotlin-dsl`
-}
-
-repositories {
-    gradlePluginPortal()
-}
-
-dependencies {
-    implementation("net.kyori", "indra-common", "3.0.1")
-    implementation("net.kyori", "indra-git", "3.0.1")
-}
-
-tasks.withType<KotlinCompile> {
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
