--- build-logic/src/main/kotlin/api.base-conventions.gradle.kts
@@ -1,27 +1,0 @@
-plugins {
-    `java-library`
-    id("net.kyori.indra")
-    id("net.kyori.indra.publishing")
-    id("net.kyori.indra.git")
-}
-
-indra {
-    github("GeyserMC", "api") {
-        ci(true)
-    }
-    mitLicense()
-
-    javaVersions {
-        target(8)
-    }
-
-    configurePublications {
-        artifactId = "${project.name}-api"
-        if (shouldAddBranchName()) {
-            version = versionWithBranchName()
-        }
-    }
-
-    publishSnapshotsTo("geysermc", "https://repo.opencollab.dev/artifactory/maven-snapshots")
-    publishReleasesTo("geysermc", "https://repo.opencollab.dev/artifactory/maven-releases")
-}
