+++ settings.gradle.kts
@@ -1,19 +1,24 @@
 @file:Suppress("UnstableApiUsage")
+
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 
 dependencyResolutionManagement {
-    repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
-    repositories {
-        maven("https://repo.opencollab.dev/main")
-        mavenCentral()
-    }
+  repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
+  repositories {
+    mavenCentral()
+    maven("https://repo.opencollab.dev/main")
+  }
 }
 
 pluginManagement {
-    repositories {
-        gradlePluginPortal()
-    }
-    includeBuild("build-logic")
+  repositories {
+    gradlePluginPortal()
+  }
+  plugins {
+    id("net.kyori.indra")
+    id("net.kyori.indra.git")
+  }
+  includeBuild("build-logic")
 }
 
 rootProject.name = "api-parent"
